requirements.txt
discord.py==2.6.4
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
pillow
gradio_client
psutil
google-genai
discord-voice
PyNaCl
mutagen
